Fix scripts/ sys.path imports; gpt2 to fp32
- scripts/predownload.py, scripts/smoke_test.py: add repo-root
  sys.path fix so `from config import ...` resolves when run as
  `python scripts/<file>.py` from repo root (same pattern already
  used by analysis/*.py).
- config.py: gpt2 dtype fp16 -> fp32, matching how every analysis/*.py
  script already loads it, and avoiding fp16 gradient-precision loss
  in the HotFlip token-substitution attack (src/hotflip.py) when run
  through main.py.

+++ config.py
@@ -6,7 +6,7 @@ MODEL_INFO = {
         'attn_implementation': 'eager',
         'eval_hidden_states': True,
         'max_batch_size': 48,
-        'dtype': 'fp16'},
+        'dtype': 'fp32'},
     'gpt2-medium': {
         'model_name': 'openai-community/gpt2-medium',
         'num_layers': 24,
